+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Modified
 - line up page
+- ProfScanner addon
 
 ## [1.2.3] - 2020-01-02
 ### Added